install-kernel: replace sh with bash
This script has been tested and seems to work as expected with bash.
Systems where sh is some other shell this script might not work at all.
By making it explicit that we are using bash, hopefully this should work
more as expected on the distros it is run on.

Signed-off-by: Alexander Alemayhu <alexander@alemayhu.com>